+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 Supported the following Fivetran API endpoints:
+- [Create account webhook](https://fivetran.com/docs/rest-api/webhooks#createaccountwebhook)
+- [Create group webhook](https://fivetran.com/docs/rest-api/webhooks#creategroupwebhook)
+- [Retrieve webhook details](https://fivetran.com/docs/rest-api/webhooks#retrievewebhookdetails)
+- [Update webhook](https://fivetran.com/docs/rest-api/webhooks#updatewebhook)
+- [Delete webhook](https://fivetran.com/docs/rest-api/webhooks#deletewebhook)
+- [Retrieve the list of webhooks](https://fivetran.com/docs/rest-api/webhooks#retrievethelistofwebhooks)
+- [Test webhook](https://fivetran.com/docs/rest-api/webhooks#testwebhook)
 - [List all roles](https://fivetran.com/docs/rest-api/roles#listallroles)
 
 ## [0.7.6](https://github.com/fivetran/go-fivetran/compare/v0.7.5...v0.7.6)